Add tax compute statment
The problem was about displaying the sales tax with two digit after the decimal point. It seems that you forget to calculate the tax. please refer to slide 36 of https://www.slideshare.net/ugurcanuzer/ch3-5910800 ; it's the original slides with the full solution. Thanks anyway ms Tarek for your effort :)